+++ src/parser/tests/lexer.rs
@@ -432,12 +432,25 @@ fn lex_set_tag() {
         "{% set a = macros::fn_call(with_args=true, name=name) %}",
         "{% set a = var | caps %}",
         "{% set a = var +1 >= 2%}",
+    ];
+
+    for i in inputs {
+        assert_lex_rule!(Rule::set_tag, i);
+    }
+}
+
+#[test]
+fn lex_set_global_tag() {
+    let inputs = vec![
+        "{% set_global a = 1 %}",
+        "{% set_global a = another_var %}",
+        "{% set_global a = another_var | filter %}",
         "{% set_global a = var +1 >= 2%}",
         "{%- set_global a = var +1 >= 2 -%}",
     ];
 
     for i in inputs {
-        assert_lex_rule!(Rule::set_tag, i);
+        assert_lex_rule!(Rule::set_global_tag, i);
     }
 }