+++ src/renderer/tests/macros.rs
@@ -191,3 +191,18 @@ fn render_macros_in_included() {
 
     assert_eq!(result.unwrap(), "my macro".to_string());
 }
+
+// https://github.com/Keats/tera/issues/255
+#[test]
+fn import_macros_into_other_macro_files() {
+    let mut tera = Tera::default();
+    tera.add_raw_templates(vec![
+        ("submacros", "{% macro test() %}Success!{% endmacro %}"),
+        ("macros", r#"{% import "submacros" as sub %}{% macro test() %}{{ sub::test() }}{% endmacro %}"#),
+        ("index", r#"{% import "macros" as macros %}{{ macros::test() }}"#),
+    ]).unwrap();
+    let result = tera.render("index", &Context::new());
+
+    assert_eq!(result.unwrap(), "Success!".to_string());
+
+}