+++ src/sort_utils.rs
@@ -33,27 +33,27 @@ pub trait GetSortKey: Ord + Sized + Clone {
 
 impl GetSortKey for OrderedF64 {
     fn get_sort_key(val: &Value) -> Result<Self> {
-        let n = val.as_f64().ok_or(format!("expected number got {}", val))?;
+        let n = val.as_f64().ok_or_else(|| format!("expected number got {}", val))?;
         OrderedF64::new(n)
     }
 }
 
 impl GetSortKey for bool {
     fn get_sort_key(val: &Value) -> Result<Self> {
-        val.as_bool().ok_or(format!("expected bool got {}", val).into())
+        val.as_bool().ok_or_else(|| format!("expected bool got {}", val).into())
     }
 }
 
 impl GetSortKey for String {
     fn get_sort_key(val: &Value) -> Result<Self> {
-        let str: Result<&str> = val.as_str().ok_or(format!("expected string got {}", val).into());
+        let str: Result<&str> = val.as_str().ok_or_else(|| format!("expected string got {}", val).into());
         Ok(str?.to_owned())
     }
 }
 
 impl GetSortKey for ArrayLen {
     fn get_sort_key(val: &Value) -> Result<Self> {
-        let arr = val.as_array().ok_or(format!("expected array got {}", val))?;
+        let arr = val.as_array().ok_or_else(|| format!("expected array got {}", val))?;
         Ok(ArrayLen(arr.len()))
     }
 }