+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 0.11.3 (unreleased)
+
+- Add a `slice` filter for arrays
+- Fix macro files importing other macro files not loading properly
+- Fix forloop container being allowed logic expressions
+- Much improved parsing error messages
+
 ## 0.11.2 (2018-02-01)
 
 - Fix regression when including templates that import macros